+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The releases of Garlmap aim to follow [semantic versioning](https://semver.org).
 
 - Module glob in favor of built-in implementation
 - Module string-similarity in favor of built-in implementation
+- Module mpv in favor of built-in implementation
 
 ### Versions