+++ CHANGELOG.md
@@ -22,8 +22,8 @@ The releases of Garlmap aim to follow [semantic versioning](https://semver.org).
 
 ### Versions
 
-- Electron 24.0.0-beta.7 (was 23.1.1)
-- Chromium 112.0.5615.39 (was 110.0.5481.104)
+- Electron 24.0.0 (was 23.1.1)
+- Chromium 112.0.5615.49 (was 110.0.5481.104)
 
 ## [1.4.0](https://github.com/Jelmerro/Garlmap/compare/1.3.1...1.4.0) - 2023-02-23